backend/requirements.txt
Django==4.1
djangorestframework
django-rest-swagger
django-filter
markdown
django-cors-headers
django-versatileimagefield
django-model-utils
djoser
requests
Pillow
python-dotenv
drf-writable-nested
django-constance[database]
psycopg2-binary
django-ckeditor
gunicorn
django-mptt
redis